+++ paramiko/primes.py
@@ -24,7 +24,6 @@ import os
 
 from paramiko import util
 from paramiko.common import byte_mask
-from paramiko.py3compat import long
 from paramiko.ssh_exception import SSHException
 
 
@@ -76,7 +75,7 @@ class ModulusPack(object):
         tries = int(tries)
         size = int(size)
         generator = int(generator)
-        modulus = long(modulus, 16)
+        modulus = int(modulus, 16)
 
         # weed out primes that aren't at least:
         # type 2 (meets basic structural requirements)
